requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
paho-mqtt==1.4.0
pkg-resources==0.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyTelegramBotAPI==3.6.6
redis==3.0.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.14
urllib3==1.24.1
